pyproject.toml
[tool.poetry]
name = "grpc-greeter"
version = "0.1.0"
description = "A gRPC service for greeting users"
authors = ["PGSch"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
grpcio = "^1.67.1"
grpcio-tools = "^1.67.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

